Split existing unit tests between plugins and they pass.
